+++ frontend/src/pages/admission-ce/create/index.tsx
@@ -1,0 +1,105 @@
+import { Button } from '@mui/material'
+import { Box } from '@mui/system'
+import { useEffect, useState } from 'react'
+import { useNavigate } from 'react-router-dom'
+import { useCreateAdmissionsMutation, useDeleteAdmissionsByIdMutation } from '../../../__data__/service/admission.api'
+import { EntityTitle } from '../../../components/entity-title'
+import { Objects } from '../../../types/api'
+import { ObjectsSelector } from '../objects-selector'
+import { EmptyAdmission } from '../../../components/admission-messages/is-empty'
+import { useDispatch, useSelector } from 'react-redux'
+import { AdmissionTechnical, setShowObjectsSelector } from '../../../__data__/states/admission-technical'
+
+export const AdmissionCreate = () => {
+  const [createAdmission, { data: createAdmissionData }] = useCreateAdmissionsMutation()
+  const [deleteAdmission] = useDeleteAdmissionsByIdMutation()
+  const dispatch = useDispatch()
+  const showObjectsSelector = useSelector(
+    (state: { admissionTechnical: AdmissionTechnical }) => state.admissionTechnical.showObjectsSelector
+  )
+  const navigate = useNavigate()
+  const [selectedObject, setSelectedObject] = useState<string[]>([])
+
+  const handleObjectSelect = (selected: Objects[]) => {
+    setSelectedObject(selected.map(obj => obj.id))
+  }
+
+  useEffect(() => {
+    if (selectedObject.length > 0) {
+      createAdmission(selectedObject)
+    }
+  }, [selectedObject])
+
+  return (
+    <>
+      {showObjectsSelector ? <ObjectsSelector onSelectObject={handleObjectSelect} /> : null}
+      <EntityTitle
+        isSwitch={false}
+        isSearchField={false}
+        customTitle={createAdmissionData?.code ? `#${createAdmissionData?.code}` : 'Код заявки'}
+      />
+      <Box
+        sx={{
+          alignItems: 'center',
+          display: 'flex',
+          justifyContent: 'center',
+          flexDirection: 'column',
+          width: '100%',
+          height: '70vh'
+        }}
+      >
+        <Box
+          sx={{
+            display: 'flex',
+            justifyContent: 'space-between',
+            alignItems: 'center',
+            width: '90%',
+            marginBottom: '30px',
+            marginTop: '8px'
+          }}
+        >
+          <Box>
+            <Button
+              variant="contained"
+              sx={{ marginRight: '14px' }}
+              onClick={() =>
+                navigate(`/admissions/${createAdmissionData?.id}/record/create`, {
+                  state: { showObjectSelector: true }
+                })
+              }
+            >
+              Добавить запись
+            </Button>
+          </Box>
+          <Box>
+            <Button variant="contained" sx={{ marginRight: '14px' }}>
+              Импортировать excel
+            </Button>
+            <Button variant="contained" disabled>
+              Архив
+            </Button>
+          </Box>
+        </Box>
+        <Box sx={{ width: '90%', height: '100%' }}>
+          <EmptyAdmission />
+        </Box>
+        <Box sx={{ width: '100%', justifyContent: 'center', display: 'flex' }}>
+          <Button
+            variant="contained"
+            sx={{ marginRight: '4%' }}
+            onClick={() => {
+              deleteAdmission(createAdmissionData ? createAdmissionData?.id : '')
+              dispatch(setShowObjectsSelector({ showObjectsSelector: true }))
+              navigate('/admissions')
+            }}
+          >
+            Отмена
+          </Button>
+          <Button variant="contained" disabled>
+            Сохранить
+          </Button>
+        </Box>
+      </Box>
+    </>
+  )
+}
