+++ frontend/src/pages/admission-ce/objects-selector/index.tsx
